refactor: 팝업 응답 형식 일치
