+++ src/main/java/be15fintomatokatchupbe/user/query/dto/response/UserSearchDto.java
@@ -1,4 +1,4 @@
-package be15fintomatokatchupbe.client.query.dto;
+package be15fintomatokatchupbe.user.query.dto.response;
 
 import lombok.Builder;
 import lombok.Getter;
@@ -7,7 +7,7 @@ import lombok.Setter;
 @Getter
 @Setter
 @Builder
-public class ClientCompanySimpleDto {
+public class UserSearchDto {
     public Long id;
     public String name;
 }
