+++ src/main/java/be15fintomatokatchupbe/user/query/mapper/UserQueryMapper.java
@@ -3,6 +3,7 @@ package be15fintomatokatchupbe.user.query.mapper;
 import be15fintomatokatchupbe.chat.query.application.dto.response.UserSimpleDto;
 import be15fintomatokatchupbe.user.query.dto.response.UserAccountQueryResponse;
 import be15fintomatokatchupbe.user.query.dto.response.UserInfluencerListDTO;
+import be15fintomatokatchupbe.user.query.dto.response.UserSearchDto;
 import org.apache.ibatis.annotations.Mapper;
 import org.apache.ibatis.annotations.Param;
 
@@ -22,5 +23,5 @@ public interface UserQueryMapper {
 
     List<Long> findUserIdsByIds(@Param("userIds") List<Long> userIds);
 
-    List<UserSimpleDto> getUserList(String keyword);
+    List<UserSearchDto> getUserList(String keyword);
 }
