+++ src/main/java/be15fintomatokatchupbe/user/query/service/UserQueryService.java
@@ -1,10 +1,7 @@
 package be15fintomatokatchupbe.user.query.service;
 
 import be15fintomatokatchupbe.chat.query.application.dto.response.UserSimpleDto;
-import be15fintomatokatchupbe.user.query.dto.response.UserAccountQueryResponse;
-import be15fintomatokatchupbe.user.query.dto.response.UserInfluencerListDTO;
-import be15fintomatokatchupbe.user.query.dto.response.UserInfluencerListResponse;
-import be15fintomatokatchupbe.user.query.dto.response.UserListResponse;
+import be15fintomatokatchupbe.user.query.dto.response.*;
 import be15fintomatokatchupbe.user.query.mapper.UserQueryMapper;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -32,7 +29,7 @@ public class UserQueryService {
     }
 
     public UserListResponse getUserList(String keyword) {
-        List<UserSimpleDto> userList = userQueryMapper.getUserList(keyword);
+        List<UserSearchDto> userList = userQueryMapper.getUserList(keyword);
 
         return UserListResponse.builder()
                 .userList(userList)
